Do refactor
