--- src/camera.rs
@@ -1,15 +1,0 @@
-//! Camera setup.
-
-use bevy::prelude::*;
-
-pub(super) fn plugin(app: &mut App) {
-    app.add_systems(Startup, spawn_camera);
-}
-
-fn spawn_camera(mut commands: Commands) {
-    commands.spawn((
-        Name::new("Camera"),
-        Camera2dBundle::default(),
-        IsDefaultUiCamera,
-    ));
-}
