+++ src/ui/mod.rs
@@ -1,4 +1,4 @@
-//! Self-contained, re-usable utilities that are not specific to this game.
+//! Re-usable UI widgets & theming.
 
 // Unused utilities and re-exports may trigger these lints undesirably.
 #![allow(dead_code, unused_imports)]
