+++ src/ui/palette.rs
@@ -8,4 +8,3 @@ pub const LABEL_TEXT: Color = Color::srgb(0.867, 0.827, 0.412);
 pub const HEADER_TEXT: Color = Color::srgb(0.867, 0.827, 0.412);
 
 pub const NODE_BACKGROUND: Color = Color::srgb(0.286, 0.478, 0.773);
-pub const BACKGROUND: Color = Color::srgb(0.58, 0.682, 0.839);
